+++ src/People.hs
@@ -7,14 +7,16 @@
 {-# LANGUAGE OverloadedStrings          #-}
 {-# LANGUAGE TemplateHaskell            #-}
 {-# LANGUAGE TypeFamilies               #-}
+{-# LANGUAGE ScopedTypeVariables        #-}
 
 module People ( PersonName(..), FirstName(..), FamilyName(..), Cognomen(..)
               , RegnalNumber(..), Sex(..), Gender(..) )
     where
 
-import Data.Aeson ( ToJSON(..), withScientific, withText )
+import Data.Aeson ( ToJSON(..), Object, withScientific, withText, withObject )
 import Data.Aeson.TH
     ( deriveJSON, defaultOptions )
+import Data.Aeson.Types ( Parser )
 import Data.Scientific ( toBoundedInteger )
 import Database.Persist.TH
 import ClassyPrelude.Yesod   as Import
@@ -25,6 +27,60 @@ data PersonName =
     | SimpleName FirstName (Maybe Cognomen)
     | RegalName FirstName FamilyName RegnalNumber (Maybe Cognomen)
     deriving (Show, Read, Eq)
+
+
+instance ToJSON PersonName where
+    toJSON (RegularName firstName familyName cognomen) =
+        object [ "Tag" .= ("RegularName" :: Text)
+               , "FirstName" .= firstName
+               , "FamilyName" .= familyName
+               , "Cognomen" .= cognomen
+               ]
+
+    toJSON (SimpleName firstName cognomen) =
+        object [ "Tag" .= ("SimpleName" :: Text)
+               , "FirstName" .= firstName
+               , "Cognomen" .= cognomen
+               ]
+
+    toJSON (RegalName firstName familyName regnalNumber cognomen) =
+        object [ "Tag" .= ("RegalName" :: Text)
+               , "FirstName" .= firstName
+               , "FamilyName" .= familyName
+               , "RegnalNumber" .= regnalNumber
+               , "Cognomen" .= cognomen
+               ]
+
+
+instance FromJSON PersonName where
+    parseJSON = withObject "person name" $ \o -> do
+        tag <- o .: "Tag"
+        name <- parseName tag o
+        return name
+
+
+-- | Parser for person names
+-- first parameter defines value constructor to use
+parseName :: Text -> Object -> Parser PersonName
+parseName "RegularName" o = do
+    firstName <- o .: "FirstName"
+    familyName <- o .: "FamilyName"
+    cognomen <- o .: "Cognomen"
+    return $ RegularName firstName familyName cognomen
+
+parseName "SimpleName" o = do
+    firstName <- o .: "FirstName"
+    cognomen <- o .: "Cognomen"
+    return $ SimpleName firstName cognomen
+
+parseName "RegalName" o = do
+    firstName <- o .: "FirstName"
+    familyName <- o .: "FamilyName"
+    regnalNumber <- o .: "RegnalNumber"
+    cognomen <- o .: "Cognomen"
+    return $ RegalName firstName familyName regnalNumber cognomen
+
+parseName _ _  = mempty
 
 
 newtype FirstName = FirstName { unFirstName :: Text }
@@ -119,6 +175,5 @@ derivePersistField "PersonName"
 derivePersistField "Sex"
 derivePersistField "Gender"
 
-$(deriveJSON defaultOptions ''PersonName)
 $(deriveJSON defaultOptions ''Sex)
 $(deriveJSON defaultOptions ''Gender)